Implement system BIOS loading, add checks for system/user binary file sizes, re-implement system/user memory maps.

+++ include/core/AddressRange.h
@@ -17,18 +17,18 @@ enum class RangeType
 class AddressRange
 {
 public:
-    AddressRange(uint16_t, uint16_t, std::optional<RangeType>);
-    uint16_t Begin() const;
-    uint16_t End() const;
+    AddressRange(size_t, size_t, std::optional<RangeType>);
+    size_t Begin() const;
+    size_t End() const;
 
 private:
-    inline void SetAllInclusive(uint16_t, uint16_t);
-    inline void SetBeginInclusive(uint16_t, uint16_t);
-    inline void SetEndInclusive(uint16_t, uint16_t);
-    inline void SetNoneInclusive(uint16_t, uint16_t);
+    inline void SetAllInclusive(size_t, size_t);
+    inline void SetBeginInclusive(size_t, size_t);
+    inline void SetEndInclusive(size_t, size_t);
+    inline void SetNoneInclusive(size_t, size_t);
 
-    uint16_t _begin;
-    uint16_t _end;
+    size_t _begin;
+    size_t _end;
 };
 
 }
