+++ include/core/interfaces/MemoryControllerInterface.h
@@ -22,11 +22,11 @@ class MemoryControllerInterface
 public:
     virtual ~MemoryControllerInterface() = default;
    
-    virtual std::variant<uint8_t, uint16_t> Read(uint16_t, interfaces::MemoryAccessType) = 0;
-    virtual void Write(std::variant<uint8_t, uint16_t>, uint16_t) = 0;
-    virtual void Load(std::unique_ptr<uint8_t*>, size_t, uint16_t, std::optional<size_t>) = 0;
+    virtual std::variant<uint8_t, uint16_t> Read(size_t, interfaces::MemoryAccessType) = 0;
+    virtual void Write(std::variant<uint8_t, uint16_t>, size_t) = 0;
+    virtual void Load(std::unique_ptr<uint8_t*>, size_t, size_t, std::optional<size_t>) = 0;
  
-    virtual void SwitchBank(uint16_t, size_t) = 0;
+    virtual void SwitchBank(size_t, size_t) = 0;
     virtual void SetMode(Mode) = 0;
 
     virtual size_t RegisterMemoryResource(std::unique_ptr<interfaces::MemoryInterface>, AddressRange, Ownership) = 0;
