+++ include/core/memory/MemoryController.h
@@ -29,7 +29,7 @@ MemoryResource;
 typedef struct ResourceIndexAndAddress_t
 {
     uint8_t ResourceIndex;
-    uint16_t LocalAddress;
+    size_t LocalAddress;
 }
 ResourceIndexAndAddress;
 
@@ -39,11 +39,11 @@ public:
     MemoryController() = default;
     virtual ~MemoryController() = default;
 
-    [[nodiscard]] std::variant<uint8_t, uint16_t> Read(uint16_t, interfaces::MemoryAccessType) override;
-    void Write(std::variant<uint8_t, uint16_t>, uint16_t) override;
-    void Load(std::unique_ptr<uint8_t*>, size_t, uint16_t, std::optional<size_t>) override;
+    [[nodiscard]] std::variant<uint8_t, uint16_t> Read(size_t, interfaces::MemoryAccessType) override;
+    void Write(std::variant<uint8_t, uint16_t>, size_t) override;
+    void Load(std::unique_ptr<uint8_t*>, size_t, size_t, std::optional<size_t>) override;
     
-    void SwitchBank(uint16_t, size_t) override;
+    void SwitchBank(size_t, size_t) override;
     void SetMode(gbxcore::Mode) override;
     gbxcore::Mode Mode();
 
@@ -56,7 +56,7 @@ private:
     inline void DetectOverlap(AddressRange);
     inline std::vector<MemoryResource>* SelectResource();
 
-    std::optional<ResourceIndexAndAddress> CalculateLocalAddress(uint16_t address);
+    std::optional<ResourceIndexAndAddress> CalculateLocalAddress(size_t address);
     std::vector<MemoryResource> _userResources; 
     std::vector<MemoryResource> _systemResources;
