+++ source/core/AddressRange.cc
@@ -3,7 +3,7 @@
 namespace gbxcore
 {
 
-AddressRange::AddressRange(uint16_t begin, uint16_t end, std::optional<RangeType> range)
+AddressRange::AddressRange(size_t begin, size_t end, std::optional<RangeType> range)
 {
     auto rangeType = range.value_or(RangeType::AllInclusive);
 
@@ -24,35 +24,35 @@ AddressRange::AddressRange(uint16_t begin, uint16_t end, std::optional<RangeType
     }
 }
 
-uint16_t AddressRange::Begin() const
+size_t AddressRange::Begin() const
 {
     return _begin;
 }
 
-uint16_t AddressRange::End() const
+size_t AddressRange::End() const
 {
     return _end;
 }
 
-inline void AddressRange::SetAllInclusive(uint16_t begin, uint16_t end)
+inline void AddressRange::SetAllInclusive(size_t begin, size_t end)
 {
     _begin = begin;
     _end = end;
 }
 
-inline void AddressRange::SetBeginInclusive(uint16_t begin, uint16_t end)
+inline void AddressRange::SetBeginInclusive(size_t begin, size_t end)
 {
     _begin = begin;
     _end = (end - 1);
 }
 
-inline void AddressRange::SetEndInclusive(uint16_t begin, uint16_t end)
+inline void AddressRange::SetEndInclusive(size_t begin, size_t end)
 {
     _begin = (begin - 1);
     _end = end;
 }
 
-inline void AddressRange::SetNoneInclusive(uint16_t begin, uint16_t end)
+inline void AddressRange::SetNoneInclusive(size_t begin, size_t end)
 {
     _begin = (begin - 1);
     _end = (end - 1);
