+++ test/test_utils/CoreTestMocksAndWrappers.h
@@ -17,11 +17,11 @@ class MemoryControllerMock : public gbxcore::interfaces::MemoryControllerInterfa
 {
 public:
     virtual ~MemoryControllerMock() = default;
-    MOCK_METHOD((std::variant<uint8_t, uint16_t>), Read, (uint16_t a, gbxcore::interfaces::MemoryAccessType b));
-    MOCK_METHOD(void, SwitchBank, (uint16_t, size_t));
+    MOCK_METHOD((std::variant<uint8_t, uint16_t>), Read, (size_t a, gbxcore::interfaces::MemoryAccessType b));
+    MOCK_METHOD(void, SwitchBank, (size_t, size_t));
     MOCK_METHOD(void, SetMode, (gbxcore::Ownership));
-    MOCK_METHOD(void, Write, ((std::variant<uint8_t, uint16_t>), uint16_t));
-    MOCK_METHOD(void, Load, ((std::unique_ptr<uint8_t*>), size_t, uint16_t, (std::optional<size_t>)));
+    MOCK_METHOD(void, Write, ((std::variant<uint8_t, uint16_t>), size_t));
+    MOCK_METHOD(void, Load, ((std::unique_ptr<uint8_t*>), size_t, size_t, (std::optional<size_t>)));
     MOCK_METHOD(size_t, RegisterMemoryResource, ((std::unique_ptr<gbxcore::interfaces::MemoryInterface>), gbxcore::AddressRange, gbxcore::Ownership));
     MOCK_METHOD(void, UnregisterMemoryResource, (size_t, gbxcore::Ownership));
 };
